Rename class of status/message command "Copy to clipboard" for consistency
