+++ app/src/main/java/net/lacolaco/smileessence/smileessence/command/message/MessageCommandCopyTextToClipboard.java
@@ -22,23 +22,23 @@
  * SOFTWARE.
  */
 
-package net.lacolaco.smileessence.command.status;
+package net.lacolaco.smileessence.command.message;
 
 import android.app.Activity;
 import android.content.ClipData;
 import android.content.ClipboardManager;
 import net.lacolaco.smileessence.R;
 import net.lacolaco.smileessence.notification.Notificator;
-import twitter4j.Status;
+import twitter4j.DirectMessage;
 
-public class StatusCommandClipboard extends StatusCommand
+public class MessageCommandCopyTextToClipboard extends MessageCommand
 {
 
     // --------------------------- CONSTRUCTORS ---------------------------
 
-    public StatusCommandClipboard(Activity activity, Status status)
+    public MessageCommandCopyTextToClipboard(Activity activity, DirectMessage message)
     {
-        super(R.id.key_command_status_clipboard, activity, status);
+        super(-1, activity, message);
     }
 
     // --------------------- GETTER / SETTER METHODS ---------------------
@@ -46,7 +46,7 @@ public class StatusCommandClipboard extends StatusCommand
     @Override
     public String getText()
     {
-        return getActivity().getString(R.string.command_status_clipboard);
+        return getActivity().getString(R.string.command_status_copy_text_to_clipboard);
     }
 
     @Override
@@ -61,7 +61,7 @@ public class StatusCommandClipboard extends StatusCommand
     public boolean execute()
     {
         ClipboardManager manager = (ClipboardManager) getActivity().getSystemService(Activity.CLIPBOARD_SERVICE);
-        manager.setPrimaryClip(ClipData.newPlainText("tweet text", getOriginalStatus().getText()));
+        manager.setPrimaryClip(ClipData.newPlainText("message text", getMessage().getText()));
         Notificator.publish(getActivity(), R.string.notice_copy_clipboard);
         return true;
     }
