+++ res/values-ja/strings.xml
@@ -213,7 +213,7 @@
     <string name="command_status_quote">引用する</string>
     <string name="command_status_quote_url">URLを引用する</string>
     <string name="command_status_quote_text">本文を引用する</string>
-    <string name="command_status_clipboard">クリップボードにコピーする</string>
+    <string name="command_status_copy_text_to_clipboard">クリップボードにコピーする</string>
     <string name="command_status_copy_url_to_clipboard">URLをクリップボードにコピーする</string>
     <string name="command_status_share">共有する</string>
     <string name="command_status_open_in_browser">ブラウザで開く</string>
